+++ packages/author-profile/__tests__/author-profile-content.web.test.js
@@ -1,4 +1,4 @@
-/* eslint-env jest */
+/* eslint-env jest, browser */
 
 import React from "react";
 import { shallow, mount } from "enzyme";
@@ -7,12 +7,9 @@ import AuthorProfileItem from "../author-profile-item";
 import AuthorProfileContent from "../author-profile-content.web.js";
 import authorProfileFixture from "../fixtures/author-profile.json";
 import articleListFixture from "../fixtures/article-list.json";
+import pagedResult from "./paged-result";
 
 test(AuthorProfileContent);
-
-afterEach(() => {
-  delete window.IntersectionObserver;
-});
 
 const results = {
   data: {
@@ -29,28 +26,34 @@ const results = {
   }
 };
 
+const intersectionObserverInstances = [];
 class FakeIntersectionObserver {
   constructor(cb) {
-    this.constructor.nodes = new Set();
-    this.constructor.cb = cb;
+    this.instanceId = intersectionObserverInstances.length;
+    intersectionObserverInstances.push({ nodes: new Set(), cb });
   }
 
   observe(node) {
     Object.defineProperty(node, "clientWidth", {
       value: 600
     });
-    this.constructor.nodes.add(node);
+    intersectionObserverInstances[this.instanceId].nodes.add(node);
   }
 
-  static dispatchAll() {
-    const entries = [...this.nodes].map((node, indx) => ({
+  static dispatchEntriesForInstance(instanceId) {
+    const instance = intersectionObserverInstances[instanceId];
+    const entries = [...instance.nodes].map((node, indx) => ({
       target: node,
       isIntersecting: indx === 0
     }));
-
-    this.cb(entries);
+    instance.cb(entries);
   }
 }
+
+afterEach(() => {
+  delete window.IntersectionObserver;
+  intersectionObserverInstances.splice(0);
+});
 
 it("renders profile articles and invoke callback on article press", done => {
   const component = shallow(
@@ -67,6 +70,7 @@ it("renders profile articles and invoke callback on article press", done => {
   );
 
   component
+    .dive()
     .find(AuthorProfileItem)
     .at(0)
     .dive()
@@ -75,15 +79,19 @@ it("renders profile articles and invoke callback on article press", done => {
     .simulate("press");
 });
 
-it("renders with an intersection observer which uses the expected options", done => {
+it("renders with an intersection observer which uses the expected options", () => {
+  // IntersectionObserver is used twice by AuthorProfileContent, once for image
+  // resizing and once for scroll tracking. We capture the opts passed so that
+  // we can assert on them later.
+  const optsSpy = jest.fn();
   window.IntersectionObserver = class {
     constructor(cb, opts) {
-      expect(opts).toMatchSnapshot();
-      done();
+      optsSpy(opts);
     }
+    observe() {} // eslint-disable-line class-methods-use-this
   };
 
-  shallow(
+  mount(
     <AuthorProfileContent
       articles={results.data.author.articles.list}
       author={authorProfileFixture.data.author}
@@ -95,6 +103,8 @@ it("renders with an intersection observer which uses the expected options", done
       onArticlePress={() => {}}
     />
   );
+
+  expect(optsSpy.mock.calls[1][0]).toMatchSnapshot();
 });
 
 it("renders a good quality image if it is visible", () => {
@@ -123,7 +133,7 @@ it("renders a good quality image if it is visible", () => {
     "//www.thetimes.co.uk/imageserver/image/%2Fmethode%2Ftimes%2Fprod%2Fweb%2Fbin%2F1b5afe88-cb0d-11e7-9ee9-e45ae7e1cdd4.jpg?crop=4252%2C2835%2C0%2C0&resize=100"
   );
 
-  window.IntersectionObserver.dispatchAll();
+  window.IntersectionObserver.dispatchEntriesForInstance(1);
 
   expect(
     component
@@ -149,7 +159,7 @@ it("renders a poor quality image if it is not visible", () => {
     />
   );
 
-  window.IntersectionObserver.dispatchAll();
+  window.IntersectionObserver.dispatchEntriesForInstance(1);
 
   expect(
     component
@@ -174,15 +184,18 @@ it("renders good quality images if there is no IntersectionObserver", () => {
   );
 
   // not ideal as this relies on the actual implementation but there's no "nice" way of setting clientWidth
-  const rn = component.instance().registerNode;
-  component.instance().registerNode = node => {
+  const authorProfileInstance = component
+    .find("AuthorProfileContent")
+    .instance();
+  const rn = authorProfileInstance.registerNode;
+  authorProfileInstance.registerNode = node => {
     if (node) {
       Object.defineProperty(node, "clientWidth", {
         value: 600
       });
     }
 
-    rn.call(component.instance(), node);
+    rn.call(authorProfileInstance, node);
   };
 
   // we have to force the render lifecycle that the lazy images rely on, in that first the nodes are registered
@@ -196,4 +209,48 @@ it("renders good quality images if there is no IntersectionObserver", () => {
   });
 
   expect(component.find("TimesImage")).toMatchSnapshot();
+});
+
+it("emits scroll tracking events for author profile content", () => {
+  window.IntersectionObserver = FakeIntersectionObserver;
+  const reporter = jest.fn();
+  const pageResults = pagedResult(0, 3);
+
+  const mountPoint = document.createElement("div");
+  document.body.appendChild(mountPoint);
+
+  mount(
+    <AuthorProfileContent
+      count={10}
+      articles={pageResults.data.author.articles.list}
+      author={authorProfileFixture.data.author}
+      slug="fiona-hamilton"
+      page={1}
+      pageSize={3}
+      imageRatio={3 / 2}
+      onTwitterLinkPress={() => {}}
+      onArticlePress={() => {}}
+    />,
+    {
+      context: {
+        tracking: {
+          analytics: reporter
+        }
+      },
+      attachTo: mountPoint
+    }
+  );
+
+  window.IntersectionObserver.dispatchEntriesForInstance(0);
+
+  expect(reporter).toHaveBeenCalledWith(
+    expect.objectContaining({
+      attrs: expect.objectContaining({
+        scrollDepth: {
+          itemNumber: 1,
+          total: 3
+        }
+      })
+    })
+  );
 });