+++ packages/author-profile/author-profile-content.js
@@ -1,10 +1,11 @@
 import React from "react";
 import { Dimensions, FlatList, StyleSheet, View } from "react-native";
+import { withTrackScrollDepth } from "@times-components/tracking";
 import AuthorProfileAuthorHead from "./author-profile-author-head";
 import AuthorProfilePagination from "./author-profile-pagination";
 import AuthorProfileItem from "./author-profile-item";
 import AuthorProfileItemSeparator from "./author-profile-item-separator";
-import propTypes from "./author-profile-content-prop-types";
+import { propTypes, defaultProps } from "./author-profile-content-prop-types";
 import { normaliseWidth } from "./utils";
 
 const styles = StyleSheet.create({
@@ -13,6 +14,11 @@ const styles = StyleSheet.create({
     paddingRight: 10
   }
 });
+
+const viewabilityConfig = {
+  viewAreaCoveragePercentThreshold: 100,
+  waitForInteraction: false
+};
 
 class AuthorProfileContent extends React.Component {
   constructor(props) {
@@ -24,6 +30,7 @@ class AuthorProfileContent extends React.Component {
       count: props.count,
       width: normaliseWidth(width)
     };
+    this.onViewableItemsChanged = this.onViewableItemsChanged.bind(this);
   }
 
   componentWillReceiveProps(nextProps) {
@@ -31,6 +38,18 @@ class AuthorProfileContent extends React.Component {
       this.setState({
         count: nextProps.count
       });
+    }
+  }
+
+  onViewableItemsChanged(info) {
+    if (info.changed) {
+      info.changed
+        .filter(viewableItem => viewableItem.isViewable)
+        .map(
+          viewableItem =>
+            this.props.onViewed &&
+            this.props.onViewed(viewableItem.item, this.props.articles)
+        );
     }
   }
 
@@ -73,7 +92,12 @@ class AuthorProfileContent extends React.Component {
             id,
             isLoading: true
           }))
-      : articles;
+      : articles.map((article, idx) => ({
+          ...article,
+          elementId: `articleList-${page}-${idx}`
+        }));
+
+    if (!articlesLoading) this.props.receiveChildList(data);
 
     return (
       <FlatList
@@ -92,6 +116,8 @@ class AuthorProfileContent extends React.Component {
           />
         )}
         initialListSize={pageSize}
+        onViewableItemsChanged={this.onViewableItemsChanged}
+        viewabilityConfig={viewabilityConfig}
         scrollRenderAheadDistance={2}
         pageSize={pageSize}
         ListHeaderComponent={
@@ -120,4 +146,5 @@ class AuthorProfileContent extends React.Component {
 }
 
 AuthorProfileContent.propTypes = propTypes;
-export default AuthorProfileContent;
+AuthorProfileContent.defaultProps = defaultProps;
+export default withTrackScrollDepth(AuthorProfileContent);