+++ packages/author-profile/CHANGELOG.md
@@ -3,6 +3,28 @@
 All notable changes to this project will be documented in this file.
 See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+<a name="0.26.0"></a>
+# [0.26.0](https://github.com/newsuk/times-components/compare/@times-components/author-profile@0.25.0...@times-components/author-profile@0.26.0) (2017-11-29)
+
+
+### Features
+
+* analytics - scroll tracking of component’s children ([#401](https://github.com/newsuk/times-components/issues/401)) ([1279f08](https://github.com/newsuk/times-components/commit/1279f08))
+
+
+
+
+<a name="0.25.0"></a>
+# [0.25.0](https://github.com/newsuk/times-components/compare/@times-components/author-profile@0.24.13...@times-components/author-profile@0.25.0) (2017-11-28)
+
+
+### Features
+
+* use summary and headline for author profile ([#417](https://github.com/newsuk/times-components/issues/417)) ([921940f](https://github.com/newsuk/times-components/commit/921940f))
+
+
+
+
 <a name="0.24.13"></a>
 ## [0.24.13](https://github.com/newsuk/times-components/compare/@times-components/author-profile@0.24.12...@times-components/author-profile@0.24.13) (2017-11-24)