+++ packages/pagination/CHANGELOG.md
@@ -3,6 +3,12 @@
 All notable changes to this project will be documented in this file.
 See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+<a name="0.10.15"></a>
+## [0.10.15](https://github.com/newsuk/times-components/compare/@times-components/pagination@0.10.14...@times-components/pagination@0.10.15) (2017-11-29)
+
+
+
+
 <a name="0.10.14"></a>
 ## [0.10.14](https://github.com/newsuk/times-components/compare/@times-components/pagination@0.10.13...@times-components/pagination@0.10.14) (2017-11-22)